Refactor: DeleteCharacter() also using logged userId

+++ Services/CharacterService/CharacterService.cs
@@ -49,12 +49,22 @@ namespace dotnet_rpg.Services.CharacterService
 
       try
       {
-        Character character = await _context.Characters.FirstAsync(c => c.Id == id);
+        Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
-        _context.Characters.Remove(character);
-        await _context.SaveChangesAsync();
+        if (character != null)
+        {
+          _context.Characters.Remove(character);
+          await _context.SaveChangesAsync();
 
-        serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+          serviceResponse.Data = _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+        }
+        else
+        {
+          serviceResponse.Success = false;
+          serviceResponse.Message = "Character not found.";
+        }
       }
       catch (Exception ex)
       {
